Make macro compile tests use bevy_compile_test_utils
