+++ crates/bevy_macros_compile_fail_tests/README.md
@@ -4,3 +4,5 @@ This crate is not part of the Bevy workspace in order to not fail `crater` tests
 The tests assert on the exact compiler errors and can easily fail for new Rust versions due to updated compiler errors (e.g. changes in spans).
 
 The `CI` workflow executes these tests on the stable rust toolchain (see [tools/ci](../../tools/ci/src/main.rs)).
+
+For information on writing tests see [bevy_compile_test_utils/README.md](../bevy_compile_test_utils/README.md).
