--- crates/bevy_macros_compile_fail_tests/tests/deref_derive.rs
@@ -1,6 +1,0 @@
-#[test]
-fn test() {
-    let t = trybuild::TestCases::new();
-    t.compile_fail("tests/deref_derive/*.fail.rs");
-    t.pass("tests/deref_derive/*.pass.rs");
-}
