+++ crates/bevy_macros_compile_fail_tests/tests/deref_derive/invalid_attribute_fail.rs
@@ -3,13 +3,16 @@ use bevy_derive::Deref;
 // Reason: `#[deref]` doesn't take any arguments
 
 #[derive(Deref)]
-struct TupleStruct(usize, #[deref()] String);
+struct TupleStruct(
+    usize,
+    #[deref()] String
+    //~^ ERROR: unexpected token
+);
 
 #[derive(Deref)]
 struct Struct {
     foo: usize,
     #[deref()]
+    //~^ ERROR: unexpected token
     bar: String,
 }
-
-fn main() {}
