--- crates/bevy_macros_compile_fail_tests/tests/deref_derive/invalid_item.fail.rs
@@ -1,9 +1,0 @@
-use bevy_derive::Deref;
-
-#[derive(Deref)]
-struct UnitStruct;
-
-#[derive(Deref)]
-enum Enum {}
-
-fn main() {}
