+++ crates/bevy_macros_compile_fail_tests/tests/deref_derive/invalid_item_fail.rs
@@ -1,0 +1,9 @@
+use bevy_derive::Deref;
+
+#[derive(Deref)]
+//~^ ERROR: cannot be derived on field-less structs
+struct UnitStruct;
+
+#[derive(Deref)]
+//~^ ERROR: can only be derived on structs
+enum Enum {}
