+++ crates/bevy_macros_compile_fail_tests/tests/deref_derive/invalid_item_fail.stderr
@@ -1,5 +1,5 @@
 error: Deref cannot be derived on field-less structs
- --> tests/deref_derive/invalid_item.fail.rs:3:10
+ --> tests/deref_derive/invalid_item_fail.rs:3:10
   |
 3 | #[derive(Deref)]
   |          ^^^^^
@@ -7,9 +7,12 @@ error: Deref cannot be derived on field-less structs
   = note: this error originates in the derive macro `Deref` (in Nightly builds, run with -Z macro-backtrace for more info)
 
 error: Deref can only be derived on structs
- --> tests/deref_derive/invalid_item.fail.rs:6:10
+ --> tests/deref_derive/invalid_item_fail.rs:7:10
   |
-6 | #[derive(Deref)]
+7 | #[derive(Deref)]
   |          ^^^^^
   |
   = note: this error originates in the derive macro `Deref` (in Nightly builds, run with -Z macro-backtrace for more info)
+
+error: aborting due to 2 previous errors
+
