+++ crates/bevy_macros_compile_fail_tests/tests/deref_derive/missing_attribute_fail.rs
@@ -1,12 +1,12 @@
 use bevy_derive::Deref;
 
 #[derive(Deref)]
+//~^ ERROR: requires one field to have
 struct TupleStruct(usize, String);
 
 #[derive(Deref)]
+//~^ ERROR: requires one field to have
 struct Struct {
     foo: usize,
     bar: String,
 }
-
-fn main() {}
