+++ crates/bevy_macros_compile_fail_tests/tests/deref_derive/missing_attribute_fail.stderr
@@ -1,5 +1,5 @@
 error: deriving Deref on multi-field structs requires one field to have the `#[deref]` attribute
- --> tests/deref_derive/missing_attribute.fail.rs:3:10
+ --> tests/deref_derive/missing_attribute_fail.rs:3:10
   |
 3 | #[derive(Deref)]
   |          ^^^^^
@@ -7,9 +7,12 @@ error: deriving Deref on multi-field structs requires one field to have the `#[d
   = note: this error originates in the derive macro `Deref` (in Nightly builds, run with -Z macro-backtrace for more info)
 
 error: deriving Deref on multi-field structs requires one field to have the `#[deref]` attribute
- --> tests/deref_derive/missing_attribute.fail.rs:6:10
+ --> tests/deref_derive/missing_attribute_fail.rs:7:10
   |
-6 | #[derive(Deref)]
+7 | #[derive(Deref)]
   |          ^^^^^
   |
   = note: this error originates in the derive macro `Deref` (in Nightly builds, run with -Z macro-backtrace for more info)
+
+error: aborting due to 2 previous errors
+
