--- crates/bevy_macros_compile_fail_tests/tests/deref_derive/multiple_attributes.fail.rs
@@ -1,14 +1,0 @@
-use bevy_derive::Deref;
-
-#[derive(Deref)]
-struct TupleStruct(#[deref] usize, #[deref] String);
-
-#[derive(Deref)]
-struct Struct {
-    #[deref]
-    foo: usize,
-    #[deref]
-    bar: String,
-}
-
-fn main() {}
