+++ crates/bevy_macros_compile_fail_tests/tests/deref_derive/multiple_attributes_fail.rs
@@ -1,0 +1,17 @@
+use bevy_derive::Deref;
+
+#[derive(Deref)]
+struct TupleStruct(
+    #[deref] usize,
+    #[deref] String
+    //~^ ERROR: can only be used on a single field
+);
+
+#[derive(Deref)]
+struct Struct {
+    #[deref]
+    foo: usize,
+    #[deref]
+    //~^ ERROR: can only be used on a single field
+    bar: String,
+}
