+++ crates/bevy_macros_compile_fail_tests/tests/deref_derive/multiple_attributes_fail.stderr
@@ -1,0 +1,14 @@
+error: `#[deref]` attribute can only be used on a single field
+ --> tests/deref_derive/multiple_attributes_fail.rs:6:5
+  |
+6 |     #[deref] String
+  |     ^^^^^^^^
+
+error: `#[deref]` attribute can only be used on a single field
+  --> tests/deref_derive/multiple_attributes_fail.rs:14:5
+   |
+14 |     #[deref]
+   |     ^^^^^^^^
+
+error: aborting due to 2 previous errors
+
