+++ crates/bevy_macros_compile_fail_tests/tests/deref_derive/single_field_pass.rs
@@ -1,3 +1,4 @@
+//@check-pass
 use bevy_derive::Deref;
 
 #[derive(Deref)]
