--- crates/bevy_macros_compile_fail_tests/tests/deref_mut_derive.rs
@@ -1,6 +1,0 @@
-#[test]
-fn test() {
-    let t = trybuild::TestCases::new();
-    t.compile_fail("tests/deref_mut_derive/*.fail.rs");
-    t.pass("tests/deref_mut_derive/*.pass.rs");
-}
