+++ crates/bevy_macros_compile_fail_tests/tests/deref_mut_derive/invalid_attribute_fail.rs
@@ -1,8 +1,14 @@
 use bevy_derive::DerefMut;
 use std::ops::Deref;
 
+// Reason: `#[deref]` doesn't take any arguments
+
 #[derive(DerefMut)]
-struct TupleStruct(usize, #[deref] String);
+struct TupleStruct(
+    usize,
+    #[deref()] String
+    //~^ ERROR: unexpected token
+);
 
 impl Deref for TupleStruct {
     type Target = String;
@@ -15,7 +21,8 @@ impl Deref for TupleStruct {
 #[derive(DerefMut)]
 struct Struct {
     foo: usize,
-    #[deref]
+    #[deref()]
+    //~^ ERROR: unexpected token
     bar: String,
 }
 
@@ -25,15 +32,4 @@ impl Deref for Struct {
     fn deref(&self) -> &Self::Target {
         &self.bar
     }
-}
-
-fn main() {
-    let mut value = TupleStruct(123, "Hello world!".to_string());
-    let _: &mut String = &mut *value;
-
-    let mut value = Struct {
-        foo: 123,
-        bar: "Hello world!".to_string(),
-    };
-    let _: &mut String = &mut *value;
 }
