+++ crates/bevy_macros_compile_fail_tests/tests/deref_mut_derive/invalid_item_fail.rs
@@ -1,9 +1,9 @@
 use bevy_derive::DerefMut;
 
 #[derive(DerefMut)]
+//~^ ERROR: cannot be derived on field-less structs
 struct UnitStruct;
 
 #[derive(DerefMut)]
+//~^ ERROR: can only be derived on structs
 enum Enum {}
-
-fn main() {}
