+++ crates/bevy_macros_compile_fail_tests/tests/deref_mut_derive/mismatched_target_type_fail.stderr
@@ -1,5 +1,5 @@
 error[E0308]: mismatched types
- --> tests/deref_mut_derive/mismatched_target_type.fail.rs:4:10
+ --> tests/deref_mut_derive/mismatched_target_type_fail.rs:4:10
   |
 4 | #[derive(DerefMut)]
   |          ^^^^^^^^
@@ -12,9 +12,9 @@ error[E0308]: mismatched types
   = note: this error originates in the derive macro `DerefMut` (in Nightly builds, run with -Z macro-backtrace for more info)
 
 error[E0308]: mismatched types
-  --> tests/deref_mut_derive/mismatched_target_type.fail.rs:15:10
+  --> tests/deref_mut_derive/mismatched_target_type_fail.rs:16:10
    |
-15 | #[derive(DerefMut)]
+16 | #[derive(DerefMut)]
    |          ^^^^^^^^
    |          |
    |          expected `&mut String`, found `&mut usize`
@@ -23,3 +23,7 @@ error[E0308]: mismatched types
    = note: expected mutable reference `&mut String`
               found mutable reference `&mut usize`
    = note: this error originates in the derive macro `DerefMut` (in Nightly builds, run with -Z macro-backtrace for more info)
+
+error: aborting due to 2 previous errors
+
+For more information about this error, try `rustc --explain E0308`.
