+++ crates/bevy_macros_compile_fail_tests/tests/deref_mut_derive/missing_attribute_fail.stderr
@@ -1,5 +1,5 @@
 error: deriving DerefMut on multi-field structs requires one field to have the `#[deref]` attribute
- --> tests/deref_mut_derive/missing_attribute.fail.rs:4:10
+ --> tests/deref_mut_derive/missing_attribute_fail.rs:4:10
   |
 4 | #[derive(DerefMut)]
   |          ^^^^^^^^
@@ -7,9 +7,12 @@ error: deriving DerefMut on multi-field structs requires one field to have the `
   = note: this error originates in the derive macro `DerefMut` (in Nightly builds, run with -Z macro-backtrace for more info)
 
 error: deriving DerefMut on multi-field structs requires one field to have the `#[deref]` attribute
-  --> tests/deref_mut_derive/missing_attribute.fail.rs:15:10
+  --> tests/deref_mut_derive/missing_attribute_fail.rs:16:10
    |
-15 | #[derive(DerefMut)]
+16 | #[derive(DerefMut)]
    |          ^^^^^^^^
    |
    = note: this error originates in the derive macro `DerefMut` (in Nightly builds, run with -Z macro-backtrace for more info)
+
+error: aborting due to 2 previous errors
+
