--- crates/bevy_macros_compile_fail_tests/tests/deref_mut_derive/missing_deref.fail.rs
@@ -1,13 +1,0 @@
-use bevy_derive::DerefMut;
-
-#[derive(DerefMut)]
-struct TupleStruct(usize, #[deref] String);
-
-#[derive(DerefMut)]
-struct Struct {
-    foo: usize,
-    #[deref]
-    bar: String,
-}
-
-fn main() {}
