--- crates/bevy_macros_compile_fail_tests/tests/deref_mut_derive/missing_deref.fail.stderr
@@ -1,33 +1,0 @@
-error[E0277]: the trait bound `TupleStruct: Deref` is not satisfied
- --> tests/deref_mut_derive/missing_deref.fail.rs:4:8
-  |
-4 | struct TupleStruct(usize, #[deref] String);
-  |        ^^^^^^^^^^^ the trait `Deref` is not implemented for `TupleStruct`
-  |
-note: required by a bound in `DerefMut`
- --> $RUST/core/src/ops/deref.rs
-
-error[E0277]: the trait bound `TupleStruct: Deref` is not satisfied
- --> tests/deref_mut_derive/missing_deref.fail.rs:3:10
-  |
-3 | #[derive(DerefMut)]
-  |          ^^^^^^^^ the trait `Deref` is not implemented for `TupleStruct`
-  |
-  = note: this error originates in the derive macro `DerefMut` (in Nightly builds, run with -Z macro-backtrace for more info)
-
-error[E0277]: the trait bound `Struct: Deref` is not satisfied
- --> tests/deref_mut_derive/missing_deref.fail.rs:7:8
-  |
-7 | struct Struct {
-  |        ^^^^^^ the trait `Deref` is not implemented for `Struct`
-  |
-note: required by a bound in `DerefMut`
- --> $RUST/core/src/ops/deref.rs
-
-error[E0277]: the trait bound `Struct: Deref` is not satisfied
- --> tests/deref_mut_derive/missing_deref.fail.rs:6:10
-  |
-6 | #[derive(DerefMut)]
-  |          ^^^^^^^^ the trait `Deref` is not implemented for `Struct`
-  |
-  = note: this error originates in the derive macro `DerefMut` (in Nightly builds, run with -Z macro-backtrace for more info)
