+++ crates/bevy_macros_compile_fail_tests/tests/deref_mut_derive/multiple_fields_pass.rs
@@ -1,19 +1,25 @@
+//@check-pass
+
 use bevy_derive::DerefMut;
 use std::ops::Deref;
 
 #[derive(DerefMut)]
-struct TupleStruct(#[deref] String);
+// The first field is never read but we want it there to check if the derive skips it.
+struct TupleStruct(#[allow(dead_code)] usize, #[deref] String);
 
 impl Deref for TupleStruct {
     type Target = String;
 
     fn deref(&self) -> &Self::Target {
-        &self.0
+        &self.1
     }
 }
 
 #[derive(DerefMut)]
 struct Struct {
+    #[allow(dead_code)]
+    // Same justification as above.
+    foo: usize,
     #[deref]
     bar: String,
 }
@@ -27,10 +33,11 @@ impl Deref for Struct {
 }
 
 fn main() {
-    let mut value = TupleStruct("Hello world!".to_string());
+    let mut value = TupleStruct(123, "Hello world!".to_string());
     let _: &mut String = &mut *value;
 
     let mut value = Struct {
+        foo: 123,
         bar: "Hello world!".to_string(),
     };
     let _: &mut String = &mut *value;
