+++ crates/bevy_macros_compile_fail_tests/tests/deref_mut_derive/single_field_pass.rs
@@ -1,21 +1,20 @@
+//@check-pass
 use bevy_derive::DerefMut;
 use std::ops::Deref;
 
 #[derive(DerefMut)]
-struct TupleStruct(#[deref] usize, #[deref] String);
+struct TupleStruct(#[deref] String);
 
 impl Deref for TupleStruct {
     type Target = String;
 
     fn deref(&self) -> &Self::Target {
-        &self.1
+        &self.0
     }
 }
 
 #[derive(DerefMut)]
 struct Struct {
-    #[deref]
-    foo: usize,
     #[deref]
     bar: String,
 }
@@ -28,4 +27,12 @@ impl Deref for Struct {
     }
 }
 
-fn main() {}
+fn main() {
+    let mut value = TupleStruct("Hello world!".to_string());
+    let _: &mut String = &mut *value;
+
+    let mut value = Struct {
+        bar: "Hello world!".to_string(),
+    };
+    let _: &mut String = &mut *value;
+}
